django/requirements.txt
asgiref==3.7.2
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
certifi==2023.11.17
cffi==1.16.0
channels==4.0.0
charset-normalizer==3.3.2
click==8.1.7
constantly==23.10.4
crispy-bootstrap4==2023.1
cryptography==41.0.7
daphne==4.0.0
Django==4.2.9
django-cors-headers==4.3.1
django-crispy-forms==2.1
djangorestframework==3.14.0
h11==0.14.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
oauthlib==3.2.2
packaging==23.2
pillow==10.2.0
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.3.0
python-dotenv==1.0.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.3.1
service-identity==24.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.9.0
Unidecode==1.3.8
urllib3==2.1.0
zope.interface==6.1